Journal Improvements

+++ src/main/org/jboss/messaging/core/journal/impl/JournalFile.java
@@ -53,7 +53,7 @@ public interface JournalFile
 	
 	int getOffset();
 	
-	long getOrderingID();
+	int getOrderingID();
 	
 	void setOffset(final int offset);
